Remove redundant proc command to list/describe/execute procs
- Resolves #22
