+++ cmd/procs/description/descriptor_test.go
@@ -1,17 +1,10 @@
 package description
 
 import (
-	"errors"
-	"fmt"
-	"net/http"
 	"testing"
 
 	"github.com/fatih/color"
-	"github.com/gojektech/proctor/daemon"
 	"github.com/gojektech/proctor/io"
-	"github.com/gojektech/proctor/proc"
-	"github.com/gojektech/proctor/proc/env"
-	"github.com/gojektech/proctor/proctord/utility"
 	"github.com/spf13/cobra"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
@@ -19,15 +12,13 @@ import (
 
 type DescribeCmdTestSuite struct {
 	suite.Suite
-	mockPrinter             *io.MockPrinter
-	mockProctorEngineClient *daemon.MockClient
-	testDescribeCmd         *cobra.Command
+	mockPrinter     *io.MockPrinter
+	testDescribeCmd *cobra.Command
 }
 
 func (s *DescribeCmdTestSuite) SetupTest() {
 	s.mockPrinter = &io.MockPrinter{}
-	s.mockProctorEngineClient = &daemon.MockClient{}
-	s.testDescribeCmd = NewCmd(s.mockPrinter, s.mockProctorEngineClient)
+	s.testDescribeCmd = NewCmd(s.mockPrinter)
 }
 
 func (s *DescribeCmdTestSuite) TestDescribeCmdUsage() {
@@ -35,76 +26,14 @@ func (s *DescribeCmdTestSuite) TestDescribeCmdUsage() {
 }
 
 func (s *DescribeCmdTestSuite) TestDescribeCmdHelp() {
-	assert.Equal(s.T(), "Describe a proc, list help for variables and constants", s.testDescribeCmd.Short)
-	assert.Equal(s.T(), "Example: proctor proc describe say-hello-world", s.testDescribeCmd.Long)
+	assert.Equal(s.T(), "[Deprecated][Correct usage: `proctor describe <proc>`]", s.testDescribeCmd.Short)
 }
 
 func (s *DescribeCmdTestSuite) TestDescribeCmdRun() {
-	arg := env.VarMetadata{
-		Name:        "arg-one",
-		Description: "arg one description",
-	}
-
-	secret := env.VarMetadata{
-		Name:        "secret-one",
-		Description: "secret one description",
-	}
-
-	doSomething := proc.Metadata{
-		Name:        "do-something",
-		Description: "does something",
-		EnvVars: env.Vars{
-			Args:    []env.VarMetadata{arg},
-			Secrets: []env.VarMetadata{secret},
-		},
-	}
-	procList := []proc.Metadata{doSomething}
-
-	s.mockProctorEngineClient.On("ListProcs").Return(procList, nil).Once()
-
-	s.mockPrinter.On("Println", fmt.Sprintf("%-40s %-100s", "Proc Name", doSomething.Name), color.Reset).Once()
-	s.mockPrinter.On("Println", fmt.Sprintf("%-40s %-100s", "Proc Description", doSomething.Description), color.Reset).Once()
-	s.mockPrinter.On("Println", fmt.Sprintf("%-40s %-100s", "Contributors", doSomething.Contributors), color.Reset).Once()
-	s.mockPrinter.On("Println", fmt.Sprintf("%-40s %-100s", "Organization", doSomething.Organization), color.Reset).Once()
-	s.mockPrinter.On("Println", "\nVariables", color.FgMagenta).Once()
-	s.mockPrinter.On("Println", fmt.Sprintf("%-40s %-100s", arg.Name, arg.Description), color.Reset).Once()
-	s.mockPrinter.On("Println", "\nConstants", color.FgMagenta).Once()
-	s.mockPrinter.On("Println", fmt.Sprintf("%-40s %-100s", secret.Name, secret.Description), color.Reset).Once()
-	s.mockPrinter.On("Println", "\nFor executing a proc, run:\nproctor proc execute <proc_name> <args_name>", color.FgGreen).Once()
+	s.mockPrinter.On("Println", "[Deprecated] Correct usage:\tproctor describe <proc>", color.FgRed).Once()
 
 	s.testDescribeCmd.Run(&cobra.Command{}, []string{"do-something"})
 
-	s.mockProctorEngineClient.AssertExpectations(s.T())
-	s.mockPrinter.AssertExpectations(s.T())
-}
-
-func (s *DescribeCmdTestSuite) TestDescribeCmdRunProctorEngineClientFailure() {
-	s.mockProctorEngineClient.On("ListProcs").Return([]proc.Metadata{}, errors.New("error")).Once()
-	s.mockPrinter.On("Println", utility.GenericDescribeCmdError, color.FgRed).Once()
-
-	s.testDescribeCmd.Run(&cobra.Command{}, []string{})
-
-	s.mockProctorEngineClient.AssertExpectations(s.T())
-	s.mockPrinter.AssertExpectations(s.T())
-}
-
-func (s *DescribeCmdTestSuite) TestDescribeCmdRunProcNotSupported() {
-	s.mockProctorEngineClient.On("ListProcs").Return([]proc.Metadata{}, nil).Once()
-	s.mockPrinter.On("Println", fmt.Sprintf("Proctor doesn't support proc: %s", "any-proc"), color.FgRed).Once()
-
-	s.testDescribeCmd.Run(&cobra.Command{}, []string{"any-proc"})
-
-	s.mockProctorEngineClient.AssertExpectations(s.T())
-	s.mockPrinter.AssertExpectations(s.T())
-}
-
-func (s *DescribeCmdTestSuite) TestDescribeCmdRunProcForUnauthorizedUser() {
-	s.mockProctorEngineClient.On("ListProcs").Return([]proc.Metadata{}, errors.New(http.StatusText(http.StatusUnauthorized))).Once()
-	s.mockPrinter.On("Println", utility.UnauthorizedError, color.FgRed).Once()
-
-	s.testDescribeCmd.Run(&cobra.Command{}, []string{"any-proc"})
-
-	s.mockProctorEngineClient.AssertExpectations(s.T())
 	s.mockPrinter.AssertExpectations(s.T())
 }
