+++ cmd/procs/execution/executioner_test.go
@@ -1,15 +1,10 @@
 package execution
 
 import (
-	"errors"
-	"fmt"
-	"net/http"
 	"testing"
 
 	"github.com/fatih/color"
-	"github.com/gojektech/proctor/daemon"
 	"github.com/gojektech/proctor/io"
-	"github.com/gojektech/proctor/proctord/utility"
 	"github.com/spf13/cobra"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
@@ -17,15 +12,13 @@ import (
 
 type ExecutionCmdTestSuite struct {
 	suite.Suite
-	mockPrinter             *io.MockPrinter
-	mockProctorEngineClient *daemon.MockClient
-	testExecutionCmd        *cobra.Command
+	mockPrinter      *io.MockPrinter
+	testExecutionCmd *cobra.Command
 }
 
 func (s *ExecutionCmdTestSuite) SetupTest() {
 	s.mockPrinter = &io.MockPrinter{}
-	s.mockProctorEngineClient = &daemon.MockClient{}
-	s.testExecutionCmd = NewCmd(s.mockPrinter, s.mockProctorEngineClient)
+	s.testExecutionCmd = NewCmd(s.mockPrinter)
 }
 
 func (s *ExecutionCmdTestSuite) TestExecutionCmdUsage() {
@@ -33,134 +26,14 @@ func (s *ExecutionCmdTestSuite) TestExecutionCmdUsage() {
 }
 
 func (s *ExecutionCmdTestSuite) TestExecutionCmdHelp() {
-	assert.Equal(s.T(), "Execute a proc with arguments given", s.testExecutionCmd.Short)
-	assert.Equal(s.T(), "Example: proctor proc execute say-hello-world SAMPLE_ARG_ONE=any SAMPLE_ARG_TWO=variable", s.testExecutionCmd.Long)
+	assert.Equal(s.T(), "[Deprecated][Correct usage: `proctor execute <proc> [args]`]", s.testExecutionCmd.Short)
 }
 
 func (s *ExecutionCmdTestSuite) TestExecutionCmd() {
-	args := []string{"say-hello-world", "SAMPLE_ARG_ONE=any", "SAMPLE_ARG_TWO=variable"}
-	procArgs := make(map[string]string)
-	procArgs["SAMPLE_ARG_ONE"] = "any"
-	procArgs["SAMPLE_ARG_TWO"] = "variable"
-
-	s.mockPrinter.On("Println", fmt.Sprintf("%-40s %-100s", "Executing Proc", "say-hello-world"), color.Reset).Once()
-	s.mockPrinter.On("Println", "With Variables", color.FgMagenta).Once()
-	s.mockPrinter.On("Println", fmt.Sprintf("%-40s %-100s", "SAMPLE_ARG_ONE", "any"), color.Reset).Once()
-	s.mockPrinter.On("Println", fmt.Sprintf("%-40s %-100s", "SAMPLE_ARG_TWO", "variable"), color.Reset).Once()
-
-	s.mockProctorEngineClient.On("ExecuteProc", "say-hello-world", procArgs).Return("executed-proc-name", nil).Once()
-
-	s.mockPrinter.On("Println", "Proc execution successful. \nStreaming logs:", color.FgGreen).Once()
-
-	s.mockProctorEngineClient.On("StreamProcLogs", "executed-proc-name").Return(nil).Once()
-	s.mockPrinter.On("Println", "Log stream of proc completed.", color.FgGreen).Once()
-
-	s.testExecutionCmd.Run(&cobra.Command{}, args)
-
-	s.mockProctorEngineClient.AssertExpectations(s.T())
-	s.mockPrinter.AssertExpectations(s.T())
-}
-
-func (s *ExecutionCmdTestSuite) TestExecutionCmdForIncorrectUsage() {
-	s.mockPrinter.On("Println", "Incorrect usage of proctor proc execute", color.FgRed).Once()
+	s.mockPrinter.On("Println", "[Deprecated] Correct usage:\tproctor execute <proc> [args]", color.FgRed).Once()
 
 	s.testExecutionCmd.Run(&cobra.Command{}, []string{})
 
-	s.mockPrinter.AssertExpectations(s.T())
-}
-
-func (s *ExecutionCmdTestSuite) TestExecutionCmdForNoProcVariables() {
-	args := []string{"say-hello-world"}
-
-	s.mockPrinter.On("Println", fmt.Sprintf("%-40s %-100s", "Executing Proc", "say-hello-world"), color.Reset).Once()
-	s.mockPrinter.On("Println", "With No Variables", color.FgRed).Once()
-
-	procArgs := make(map[string]string)
-	s.mockProctorEngineClient.On("ExecuteProc", "say-hello-world", procArgs).Return("executed-proc-name", nil).Once()
-
-	s.mockPrinter.On("Println", "Proc execution successful. \nStreaming logs:", color.FgGreen).Once()
-
-	s.mockProctorEngineClient.On("StreamProcLogs", "executed-proc-name").Return(nil).Once()
-	s.mockPrinter.On("Println", "Log stream of proc completed.", color.FgGreen).Once()
-
-	s.testExecutionCmd.Run(&cobra.Command{}, args)
-
-	s.mockProctorEngineClient.AssertExpectations(s.T())
-	s.mockPrinter.AssertExpectations(s.T())
-}
-
-func (s *ExecutionCmdTestSuite) TestExecutionCmdForIncorrectVariableFormat() {
-	args := []string{"say-hello-world", "incorrect-format"}
-
-	s.mockPrinter.On("Println", fmt.Sprintf("%-40s %-100s", "Executing Proc", "say-hello-world"), color.Reset).Once()
-	s.mockPrinter.On("Println", "With Variables", color.FgMagenta).Once()
-	s.mockPrinter.On("Println", fmt.Sprintf("%-40s %-100s", "\nIncorrect variable format\n", "incorrect-format"), color.FgRed).Once()
-
-	procArgs := make(map[string]string)
-	s.mockProctorEngineClient.On("ExecuteProc", "say-hello-world", procArgs).Return("executed-proc-name", nil).Once()
-
-	s.mockPrinter.On("Println", "Proc execution successful. \nStreaming logs:", color.FgGreen).Once()
-
-	s.mockProctorEngineClient.On("StreamProcLogs", "executed-proc-name").Return(nil).Once()
-	s.mockPrinter.On("Println", "Log stream of proc completed.", color.FgGreen).Once()
-
-	s.testExecutionCmd.Run(&cobra.Command{}, args)
-
-	s.mockProctorEngineClient.AssertExpectations(s.T())
-	s.mockPrinter.AssertExpectations(s.T())
-}
-
-func (s *ExecutionCmdTestSuite) TestExecutionCmdForProctorEngineExecutionFailure() {
-	args := []string{"say-hello-world"}
-
-	s.mockPrinter.On("Println", fmt.Sprintf("%-40s %-100s", "Executing Proc", "say-hello-world"), color.Reset).Once()
-	s.mockPrinter.On("Println", "With No Variables", color.FgRed).Once()
-
-	procArgs := make(map[string]string)
-	s.mockProctorEngineClient.On("ExecuteProc", "say-hello-world", procArgs).Return("", errors.New("error")).Once()
-
-	s.mockPrinter.On("Println", utility.GenericProcCmdError, color.FgRed).Once()
-
-	s.testExecutionCmd.Run(&cobra.Command{}, args)
-
-	s.mockProctorEngineClient.AssertExpectations(s.T())
-	s.mockPrinter.AssertExpectations(s.T())
-}
-
-func (s *ExecutionCmdTestSuite) TestExecutionCmdForProctorEngineExecutionForUnauthorizedUser() {
-	args := []string{"say-hello-world"}
-
-	s.mockPrinter.On("Println", fmt.Sprintf("%-40s %-100s", "Executing Proc", "say-hello-world"), color.Reset).Once()
-	s.mockPrinter.On("Println", "With No Variables", color.FgRed).Once()
-
-	procArgs := make(map[string]string)
-	s.mockProctorEngineClient.On("ExecuteProc", "say-hello-world", procArgs).Return("", errors.New(http.StatusText(http.StatusUnauthorized))).Once()
-
-	s.mockPrinter.On("Println", utility.UnauthorizedError, color.FgRed).Once()
-
-	s.testExecutionCmd.Run(&cobra.Command{}, args)
-
-	s.mockProctorEngineClient.AssertExpectations(s.T())
-	s.mockPrinter.AssertExpectations(s.T())
-}
-
-func (s *ExecutionCmdTestSuite) TestExecutionCmdForProctorEngineLogStreamingFailure() {
-	args := []string{"say-hello-world"}
-
-	s.mockPrinter.On("Println", fmt.Sprintf("%-40s %-100s", "Executing Proc", "say-hello-world"), color.Reset).Once()
-	s.mockPrinter.On("Println", "With No Variables", color.FgRed).Once()
-
-	procArgs := make(map[string]string)
-	s.mockProctorEngineClient.On("ExecuteProc", "say-hello-world", procArgs).Return("executed-proc-name", nil).Once()
-
-	s.mockPrinter.On("Println", "Proc execution successful. \nStreaming logs:", color.FgGreen).Once()
-
-	s.mockProctorEngineClient.On("StreamProcLogs", "executed-proc-name").Return(errors.New("error")).Once()
-	s.mockPrinter.On("Println", "Error Streaming Logs", color.FgRed).Once()
-
-	s.testExecutionCmd.Run(&cobra.Command{}, args)
-
-	s.mockProctorEngineClient.AssertExpectations(s.T())
 	s.mockPrinter.AssertExpectations(s.T())
 }
