+++ cmd/procs/proc.go
@@ -4,25 +4,23 @@ import (
 	"github.com/gojektech/proctor/cmd/procs/description"
 	"github.com/gojektech/proctor/cmd/procs/execution"
 	"github.com/gojektech/proctor/cmd/procs/list"
-	"github.com/gojektech/proctor/daemon"
 	"github.com/gojektech/proctor/io"
 	"github.com/spf13/cobra"
 )
 
-func NewCmd(printer io.Printer, proctorEngineClient daemon.Client) *cobra.Command {
+func NewCmd(printer io.Printer) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "proc",
-		Short: "Interact with proctor procs",
-		Long:  `Example: proctor proc <command>`,
+		Short: "[Deprecated][Correct Usage: `proctor list/describe/execute`]",
 	}
 
-	descriptionCmd := description.NewCmd(printer, proctorEngineClient)
+	descriptionCmd := description.NewCmd(printer)
 	cmd.AddCommand(descriptionCmd)
 
-	executionCmd := execution.NewCmd(printer, proctorEngineClient)
+	executionCmd := execution.NewCmd(printer)
 	cmd.AddCommand(executionCmd)
 
-	listCmd := list.NewCmd(printer, proctorEngineClient)
+	listCmd := list.NewCmd(printer)
 	cmd.AddCommand(listCmd)
 
 	return cmd
