+++ proctord/.env.test
@@ -1,21 +1,22 @@
-export PROCTOR_KUBE_CONFIG="out-of-cluster"
-export PROCTOR_LOG_LEVEL="debug"
-export PROCTOR_APP_PORT="5000"
-export PROCTOR_DEFAULT_NAMESPACE="default"
-export PROCTOR_REDIS_ADDRESS="localhost:6379"
-export PROCTOR_REDIS_MAX_ACTIVE_CONNECTIONS="10"
-export PROCTOR_KUBE_JOB_ACTIVE_DEADLINE_SECONDS="60"
-export PROCTOR_KUBE_JOB_RETRIES="0"
-export PROCTOR_LOGS_STREAM_READ_BUFFER_SIZE="140"
-export PROCTOR_LOGS_STREAM_WRITE_BUFFER_SIZE="4096"
-export PROCTOR_KUBE_CLUSTER_HOST_NAME="localhost:8001"
-export PROCTOR_KUBE_POD_LIST_WAIT_TIME="5"
-export PROCTOR_KUBE_CA_CERT_ENCODED="LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCmNlcnRpZmljYXRlCi0tLS0tRU5EIENFUlRJRklDQVRFLS0tLS0K"
-export PROCTOR_KUBE_BASIC_AUTH="YWRtaW46cGFzc3dvcmQK"
-export PROCTOR_POSTGRES_USER="postgres"
-export PROCTOR_POSTGRES_PASSWORD=""
-export PROCTOR_POSTGRES_HOST="localhost"
-export PROCTOR_POSTGRES_PORT="5432"
-export PROCTOR_POSTGRES_DATABASE="proctord_test"
-export PROCTOR_POSTGRES_MAX_CONNECTIONS="50"
-export PROCTOR_POSTGRES_CONNECTIONS_MAX_LIFETIME="30"
+ENVIRONMENT=test
+PROCTOR_KUBE_CONFIG=out-of-cluster
+PROCTOR_LOG_LEVEL=debug
+PROCTOR_APP_PORT=5000
+PROCTOR_DEFAULT_NAMESPACE=default
+PROCTOR_REDIS_ADDRESS=localhost:6379
+PROCTOR_REDIS_MAX_ACTIVE_CONNECTIONS=10
+PROCTOR_KUBE_JOB_ACTIVE_DEADLINE_SECONDS=60
+PROCTOR_KUBE_JOB_RETRIES=0
+PROCTOR_LOGS_STREAM_READ_BUFFER_SIZE=140
+PROCTOR_LOGS_STREAM_WRITE_BUFFER_SIZE=4096
+PROCTOR_KUBE_CLUSTER_HOST_NAME=localhost:8001
+PROCTOR_KUBE_POD_LIST_WAIT_TIME=5
+PROCTOR_KUBE_CA_CERT_ENCODED=LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCmNlcnRpZmljYXRlCi0tLS0tRU5EIENFUlRJRklDQVRFLS0tLS0K
+PROCTOR_KUBE_BASIC_AUTH=YWRtaW46cGFzc3dvcmQK
+PROCTOR_POSTGRES_USER=postgres
+PROCTOR_POSTGRES_PASSWORD=
+PROCTOR_POSTGRES_HOST=localhost
+PROCTOR_POSTGRES_PORT=5432
+PROCTOR_POSTGRES_DATABASE=proctord_test
+PROCTOR_POSTGRES_MAX_CONNECTIONS=50
+PROCTOR_POSTGRES_CONNECTIONS_MAX_LIFETIME=30
