+++ proctord/Makefile
@@ -1,24 +1,31 @@
+#!make
+include .env.test
+export $(shell sed 's/=.*//' .env.test)
+
 APP=proctord
-DB_USER=$(PROCTOR_POSTGRES_USER)
-DB_NAME=$(PROCTOR_POSTGRES_DATABASE)
 APP_EXECUTABLE=./$(APP)
 
 build-deps:
 	glide install
 
-build: build-deps
-	go build
+build:
+	go build -o $(APP)
 
 test:
-	source .env.test && go test $(shell glide novendor)
+	go test -race -cover $(shell glide novendor)
+
+test-verbose:
+	go test -race -cover $(shell glide novendor) -v
 
 db.setup: db.create db.migrate
 
 db.create:
-	createdb -O$(DB_USER) -Eutf8 $(DB_NAME)
+	PGPASSWORD=$(PROCTOR_POSTGRES_PASSWORD) psql -h $(PROCTOR_POSTGRES_HOST) -p $(PROCTOR_POSTGRES_PORT) -c 'create database $(PROCTOR_POSTGRES_DATABASE);' -U $(PROCTOR_POSTGRES_USER)
 
 db.migrate: $(APP_EXECUTABLE)
 	$(APP_EXECUTABLE) migrate
 
-copy-config:
-	cp .env.sample .env
+db.teardown:
+	PGPASSWORD=$(PROCTOR_POSTGRES_PASSWORD) psql -h $(PROCTOR_POSTGRES_HOST) -p $(PROCTOR_POSTGRES_PORT) -c 'drop database $(PROCTOR_POSTGRES_DATABASE);' -U $(PROCTOR_POSTGRES_USER)
+
+setup-test-teardown: build-deps build db.setup test-verbose db.teardown
